More stuff

+++ Life Logger/LogTableViewController.swift
@@ -22,7 +22,7 @@ class LogTableViewController: UITableViewController {
         // self.clearsSelectionOnViewWillAppear = false
 
         // Uncomment the following line to display an Edit button in the navigation bar for this view controller.
-        // self.navigationItem.rightBarButtonItem = self.editButtonItem()
+         self.navigationItem.rightBarButtonItem = self.editButtonItem
         
         logs = DatabaseController.loadLogs()
     }
@@ -95,6 +95,12 @@ class LogTableViewController: UITableViewController {
     // Override to support conditional editing of the table view.
     override func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool {
         // Return false if you do not want the specified item to be editable.
+        let log = logs[indexPath.row]
+        
+        if log.dateEnded == nil {
+            return false
+        }
+        
         return true
     }
 
@@ -103,13 +109,16 @@ class LogTableViewController: UITableViewController {
         if editingStyle == .delete {
             // Delete the row from the data source
             let logToDelete = logs[indexPath.row]
-            let context = DatabaseController.getContext()
-            context.delete(logToDelete)
             
-            logs.remove(at: indexPath.row)
-            DatabaseController.saveContext()
-            
-            tableView.deleteRows(at: [indexPath], with: .fade)
+            if logToDelete.dateEnded != nil {
+                let context = DatabaseController.getContext()
+                context.delete(logToDelete)
+                
+                logs.remove(at: indexPath.row)
+                DatabaseController.saveContext()
+                
+                tableView.deleteRows(at: [indexPath], with: .fade)
+            }
         } else if editingStyle == .insert {
             // Create a new instance of the appropriate class, insert it into the array, and add a new row to the table view
         }
